+++ mechanism_manager/include/mechanism_manager/mechanism_manager_interface.h
@@ -93,11 +93,11 @@ class MechanismManagerInterface
     double GetPhase(const int idx);
     double GetScale(const int idx);
     void GetVmMode(std::string& mode);
-    void GetMergeThreshold(int& merge_th);
+    void GetMergeThreshold(double& merge_th);
 
     /// Sets
     void SetVmMode(const scale_mode_t mode);
-    void SetMergeThreshold(int merge_th);
+    void SetMergeThreshold(double merge_th);
 
   protected:
